[EA-196] Transactional channel: Transaction details alliance
