+++ src/app/core/utils/datatables.ts
@@ -177,5 +177,8 @@ export class Datatables {
         return false;
       }
   }
+  public formatTableName(tableName: string): string {
+    return tableName.split('_').map((name: string) => name.charAt(0).toUpperCase() + name.slice(1)).join(' ');
+  }
 
 }
