+++ src/app/loans/common-resolvers/loan-transaction-datatable.resolver.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { LoanTransactionDatatableResolver } from './loan-transaction-datatable.resolver';
+
+describe('LoanTransactionDatatableResolver', () => {
+  let resolver: LoanTransactionDatatableResolver;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    resolver = TestBed.inject(LoanTransactionDatatableResolver);
+  });
+
+  it('should be created', () => {
+    expect(resolver).toBeTruthy();
+  });
+});
