+++ src/app/loans/common-resolvers/loan-transaction-datatables.resolver.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { LoanTransactionDatatablesResolver } from './loan-transaction-datatables.resolver';
+
+describe('LoanTransactionDatatablesResolver', () => {
+  let resolver: LoanTransactionDatatablesResolver;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    resolver = TestBed.inject(LoanTransactionDatatablesResolver);
+  });
+
+  it('should be created', () => {
+    expect(resolver).toBeTruthy();
+  });
+});
