+++ src/app/loans/loans-view/transactions/view-transaction/datatable-transaction-tab/datatable-transaction-tab.component.html
@@ -1,0 +1,9 @@
+<div class="tab-container mat-typography">
+  <mifosx-entity-datatable-tab
+    entityType="Loan Account Transaction"
+    [entityId]="entityId"
+    [multiRowDatatableFlag]="multiRowDatatableFlag"
+    [entityDatatable]="entityDatatable"
+  >
+  </mifosx-entity-datatable-tab>
+</div>
