+++ src/app/loans/loans-view/transactions/view-transaction/datatable-transaction-tab/datatable-transaction-tab.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { DatatableTransactionTabComponent } from './datatable-transaction-tab.component';
+
+describe('DatatableTransactionTabComponent', () => {
+  let component: DatatableTransactionTabComponent;
+  let fixture: ComponentFixture<DatatableTransactionTabComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ DatatableTransactionTabComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(DatatableTransactionTabComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
