+++ src/app/loans/loans-view/transactions/view-transaction/loan-transaction-general-tab/loan-transaction-general-tab.component.html
@@ -1,0 +1,188 @@
+<div fxLayoutAlign="end" class="container m-b-20 transaction-buttons" fxLayoutGap="2%"
+     *ngIf="!(transactionData.type.contra)">
+  <span *mifosxHasPermission="'ADJUST_LOAN'">
+    <button mat-raised-button color="primary" [routerLink]="'edit'" [disabled]="!allowEdition">
+      <fa-icon icon="edit" class="m-r-10"></fa-icon>{{ 'labels.buttons.Edit' | translate}}
+    </button>
+
+    <button mat-raised-button color="warn" [routerLink]="'undo'" [disabled]="!allowUndo">
+      <fa-icon icon="undo" class="m-r-10"></fa-icon>{{ 'labels.buttons.Undo' | translate}}
+    </button>
+  </span>
+  <span *mifosxHasPermission="'CHARGEBACK_LOAN'">
+    <button mat-raised-button color="warn" [disabled]="!allowChargeback" (click)="chargebackTransaction()">
+      <fa-icon icon="undo" class="m-r-10"></fa-icon>{{ 'labels.buttons.Chargeback' | translate}}
+    </button>
+  </span>
+</div>
+
+<div class="container">
+
+  <mat-card>
+
+    <mat-card-content>
+
+      <div fxLayout="row wrap" class="content">
+        <div fxFlex="100%" [ngClass]="loanTransactionColor()">
+        </div>
+
+        <div fxFlex="50%" class="mat-body-strong">
+          {{ 'labels.inputs.Transaction Id' | translate}}
+        </div>
+
+        <div fxFlex="50%">
+          {{ transactionData.id }}
+        </div>
+
+        <div fxFlex="50%" class="mat-body-strong" *ngIf="transactionData.toLoanTransactionRelation">
+          {{ 'Repayment Transaction' | translate}}
+        </div>
+
+        <div fxFlex="50%" *ngIf="transactionData.toLoanTransactionRelation">
+          <a [href]="loanTransactionRelatedLink(transactionData.toLoanTransactionRelation.fromLoanTransaction)">
+            {{ transactionData.toLoanTransactionRelation.fromLoanTransaction }}
+          </a>
+        </div>
+
+        <div fxFlex="50%" class="mat-body-strong">
+          {{ 'labels.inputs.Type' | translate}}
+        </div>
+
+        <div fxFlex="50%">
+          {{ transactionData.type.code | translate }}
+        </div>
+
+        <div fxFlex="50%" class="mat-body-strong">
+          {{ 'labels.inputs.Transaction Date' | translate}}
+        </div>
+
+        <div fxFlex="50%">
+          {{ transactionData.date | dateFormat }}
+        </div>
+
+        <div fxFlex="50%" class="mat-body-strong">
+          {{ 'labels.inputs.Currency' | translate}}
+        </div>
+
+        <div fxFlex="50%">
+          {{ transactionData.currency.name }}
+        </div>
+
+        <div fxFlex="50%" class="mat-body-strong">
+          {{ 'labels.inputs.Amount' | translate}}
+        </div>
+
+        <div fxFlex="50%">
+          {{ transactionData.amount | currency:transactionData.currency.code:'symbol-narrow':format:locale }}
+        </div>
+
+        <div fxFlex="50%" class="mat-body-strong" *ngIf="transactionData.type.disbursement">
+          {{ 'labels.inputs.Disbursement Fees' | translate}}
+        </div>
+
+        <div fxFlex="50%" *ngIf="transactionData.type.disbursement">
+          {{ transactionData.disbursementFeesSummary }}
+        </div>
+
+        <div fxFlex="50%" class="mat-body-strong" *ngIf="transactionData.type.disbursement">
+          {{ 'labels.inputs.Total Disbursement Fees' | translate}}
+        </div>
+
+        <div fxFlex="50%" *ngIf="transactionData.type.disbursement">
+          {{ transactionData.totalDisbursementFeesAmount | currency:transactionData.currency.code:'symbol-narrow':format:locale }}
+        </div>
+
+        <div fxFlex="50%" class="mat-body-strong" *ngIf="transactionData.type.disbursement">
+          {{ 'labels.inputs.Net Disbursal Amount' | translate}}
+        </div>
+
+        <div fxFlex="50%" *ngIf="transactionData.type.disbursement">
+          {{ transactionData.netDisbursalAmount | currency:transactionData.currency.code:'symbol-narrow':format:locale }}
+        </div>
+
+        <div fxFlex="50%" class="mat-body-strong" *ngIf="channelName">
+          {{ 'labels.inputs.Channel' | translate}}
+        </div>
+
+        <div fxFlex="50%" *ngIf="channelName">
+          {{ channelName }}
+        </div>
+
+        <div fxFlex="50%" class="mat-body-strong" *ngIf="bankName">
+          {{ 'labels.inputs.Bank Name' | translate}}
+        </div>
+
+        <div fxFlex="50%" *ngIf="transactionData.paymentDetailData?.bankName">
+          {{ transactionData.paymentDetailData.bankName }}
+        </div>
+
+        <div fxFlex="50%" class="mat-body-strong" *ngIf="pointOfSalesName">
+          {{ 'labels.inputs.pointOfSalesName' | translate}}
+        </div>
+
+        <div fxFlex="50%" *ngIf="pointOfSalesName">
+          {{ pointOfSalesName }}
+        </div>
+
+        <div fxFlex="50%" class="mat-body-strong" *ngIf="transactionData.externalId">
+          {{ 'labels.inputs.External Id' | translate}}
+        </div>
+
+        <div fxFlex="50%" *ngIf="transactionData.externalId">
+          <mifosx-external-identifier externalId="{{transactionData.externalId}}" completed="true"></mifosx-external-identifier>
+        </div>
+        <div fxFlex="50%" class="mat-body-strong">
+          {{ 'labels.inputs.cashierName' | translate}}
+        </div>
+
+        <div fxFlex="50%">
+          {{ lastModifiedByFullName() }}
+        </div>
+
+        <mat-divider *ngIf="existTransactionRelations" [inset]="true"></mat-divider>
+
+        <div fxFlex="100%" class="mat-body-strong" *ngIf="existTransactionRelations">
+          {{ 'Linked Transactions' | translate}}
+          <br>
+
+          <table mat-table [dataSource]="transactionRelations">
+            <ng-container matColumnDef="relationType">
+              <th mat-header-cell *matHeaderCellDef> {{"labels.inputs.Relation Type" | translate}} </th>
+              <td mat-cell *matCellDef="let transactionRel" [matTooltip]="transactionRel.paymentType">
+                {{ transactionRel.relationType | translate}}
+              </td>
+            </ng-container>
+
+            <ng-container matColumnDef="toTransaction">
+              <th mat-header-cell *matHeaderCellDef> {{"labels.inputs.Transaction" | translate}} </th>
+              <td mat-cell *matCellDef="let transactionRel">
+                <a [href]="loanTransactionRelatedLink(transactionRel.toLoanTransaction)">
+                  {{ transactionRel.toLoanTransaction }}
+                </a>
+              </td>
+            </ng-container>
+
+            <ng-container matColumnDef="amount">
+              <th mat-header-cell *matHeaderCellDef> {{"labels.inputs.Amount" | translate}} </th>
+              <td mat-cell *matCellDef="let transactionRel">
+                {{ transactionRel.amount | currency:transactionData.currency.code:'symbol-narrow':format:locale }}
+              </td>
+            </ng-container>
+            <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+            <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+          </table>
+        </div>
+
+        <mifosx-transaction-payment-detail *ngIf="transactionData.paymentDetailData"
+                                           [paymentDetailData]="transactionData.paymentDetailData">
+        </mifosx-transaction-payment-detail>
+      </div>
+
+      <div fxLayout="row" fxLayoutAlign="center" fxLayoutGap="2%" fxLayout.lt-md="column">
+        <button type="button" mat-raised-button color="primary" [routerLink]="['../../']">{{ 'Back' | translate}}</button>
+      </div>
+    </mat-card-content>
+
+  </mat-card>
+
+</div>
