+++ src/app/loans/loans-view/transactions/view-transaction/loan-transaction-general-tab/loan-transaction-general-tab.component.scss
@@ -1,0 +1,39 @@
+@import "assets/styles/helper";
+
+.container {
+  max-width: 37rem;
+  .content {
+    div {
+      margin: 1rem 0;
+      word-wrap: break-word;
+    }
+
+    mifosx-transaction-payment-detail {
+      width: 100%;
+      margin-bottom: 20px;
+    }
+  }
+}
+
+.transaction-buttons {
+  margin-top: 3%;
+}
+
+.loan-relation {
+  background-color: $mid-grey;
+}
+
+.active {
+  background-color: $status-active;
+  height: 4px;
+}
+
+.linked {
+  background-color: $status-approved;
+  height: 4px;
+}
+
+.undo {
+  background-color: $status-active-overdue;
+  height: 4px;
+}
